refactor: move audio processing into dsp
